기능추가: login된 사용자만 수정/삭제 가능 && Post에 작성자의 id가 표시되도록 수정

+++ posts/views.py
@@ -28,7 +28,11 @@ def list(request):
 
 @require_POST
 def delete(request, post_num):
-    post = Post.objects.get(id=post_num)
+    post = get_object_or_404(Post, pk=post_num)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+    
     post.delete()
     
     return redirect('posts:list')
@@ -36,6 +40,10 @@ def delete(request, post_num):
 
 def update(request, post_num):
     post = get_object_or_404(Post, pk=post_num)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+    
     if request.method == 'GET':
         form = PostForm(instance=post)
